requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
textblob>=0.15.3
tweepy>=4.0.0
streamlit>=1.0.0
transformers>=4.0.0
python-dotenv>=0.19.0  # For managing API keys securely